refactor: move some operations to single files
- moved lastStep function
- moved quickstart operation logics to single file
