+++ index.js
@@ -3,18 +3,17 @@
 var chalk = require('chalk')
 var clear = require('clear')
 var figlet = require('figlet')
-var fs = require('fs')
-var path = require('path')
 var inquirer = require('inquirer')
-var ghdownload = require('github-download')
-var replace = require('./replace')
+var parseArgs = require('minimist')
+
 var api = require('./api')
+const lastStep = require('./lib/last-step')
+
 var pushComponents = require('./tasks/push-components')
 var pullComponents = require('./tasks/pull-components')
 var deleteTemplates = require('./tasks/delete-templates')
 var scaffold = require('./tasks/scaffold')
-var opn = require('opn')
-var parseArgs = require('minimist')
+const quickstart = require('./tasks/quickstart')
 
 clear()
 console.log(chalk.cyan(figlet.textSync('Storyblok')))
@@ -27,6 +26,7 @@ var subcommand = 'quickstart'
 var cliAttribute = ''
 var argv = parseArgs(process.argv.slice(2))
 
+// TODO: refactor to use a function
 if (typeof argv._[0] != 'undefined') {
   subcommand = argv._[0]
 
@@ -38,6 +38,7 @@ if (typeof argv._[0] != 'undefined') {
 var questions = []
 var email = ''
 
+// TODO: refactor to use a function
 if (['select'].indexOf(subcommand) > -1) {
   console.log()
   console.log()
@@ -286,183 +287,10 @@ inquirer.prompt(questions).then(function (answers) {
     answers.type = 'quickstart'
   }
 
-  var lastStep = function () {
-    var regex = /\[(.*?)\]/;
-    var gitRepo = '';
-
-    if (answers.type == 'Theme (Storyrenderer/Hosted)' || answers.type == 'Boilerplate (Selfhosted)') {
-      if (answers.custom_theme_url) {
-        gitRepo = answers.custom_theme_url + '.git'
-      } else {
-        gitRepo = regex.exec(answers.theme)[1] + '.git'
-      }
-    }
-
-    if (answers.type == 'Fieldtype') {
-      gitRepo = 'https://github.com/storyblok/storyblok-fieldtype.git'
-    }
-
-    if (answers.type == 'quickstart') {
-      gitRepo = 'https://github.com/storyblok/quickstart.git'
-    }
-
-    var outputDir = './' + answers.name
-
-    console.log(chalk.green('✓') + ' - The github repository ' + gitRepo + ' will be cloned now...')
-    ghdownload(gitRepo, outputDir)
-      .on('error', function (err) {
-        if (err.code == 'ENOTEMPTY') {
-          console.log()
-          console.log(chalk.red('  Oh Snap! It seems that you already have a project with the name: ' + answers.name))
-          console.log()
-        } else {
-          console.log()
-          console.log(chalk.red('  We never had this kind of issue - Sorry for that!'))
-          console.log(chalk.red('  Could you send us the error below as a stackoverflow question?'))
-          console.log(chalk.red('  That would be great! :)'))
-          console.log(chalk.red('  Don\'t forget to mark it with the tag `storyblok` so will can find it.'))
-          console.log()
-          console.error(err)
-          process.exit(0)
-        }
-      })
-      .on('end', function () {
-        var finalStep = 'gulp'
-        
-        if (answers.type == 'Fieldtype' || answers.type == 'quickstart') {
-          finalStep = 'npm run dev'
-        }
-
-        console.log(chalk.green('✓') + ' - ' + chalk.white('Your Storyblok project is ready for you!'))
-        console.log()
-        console.log(chalk.white('  Execute the following command to start Storyblok:'))
-        console.log(chalk.cyan('  cd ./' + answers.name + ' && npm install && ' + finalStep))
-        console.log()
-        console.log(chalk.white('  If you need more help, just try asking a question on stackoverflow'))
-        console.log(chalk.white('  with the [storyblok] tag or live-chat with us on www.storyblok.com'))
-        console.log()
-
-        if (answers.type == 'Theme (Storyrenderer/Hosted)' || answers.type == 'quickstart') {
-          fs.renameSync(outputDir + '/_token.js', outputDir + '/token.js')
-
-          if (answers.themeToken) {
-            replace(outputDir + '/token.js', { 'INSERT_TOKEN': answers.themeToken })
-          }
-
-          var spaceConfig = {}
-
-          if (answers.spaceId) {
-            spaceConfig['INSERT_SPACE_ID'] = answers.spaceId
-          }
-
-          if (answers.spaceDomain) {
-            spaceConfig['INSERT_YOUR_DOMAIN'] = answers.spaceDomain
-          }
-
-          if (answers.loginToken) {
-            spaceConfig['TEMP_QUICKSTART_TOKEN'] = answers.loginToken
-          }
-
-          replace(outputDir + '/config.js', spaceConfig)
-        }
-      })
-  }
-
   switch (subcommand) {
     case 'quickstart':
       var spaceId = argv.space
-
-      if (typeof spaceId !== 'undefined') {
-
-        console.log('  Your project will be initialized now...')
-
-        api.put('spaces/' + spaceId + '/', {
-          space: {
-            environments: [{ name: 'Dev', location: 'http://localhost:4440/' }]
-          }
-        }, (space_res) => {
-          if (space_res.status == 200) {
-            answers.name = space_res.body.space.name.replace(/\s+/g, '-').toLowerCase()
-            console.log(chalk.green('✓') + ' - Space ' + answers.name + ' updated with dev environment in Storyblok')
-
-            api.post('tokens', {}, (tokens_res) => {
-              console.log(chalk.green('✓') + ' - Configuration for your Space was loaded')
-              answers.loginToken = tokens_res.body.key
-
-              api.setSpaceId(space_res.body.space.id)
-              api.get('api_keys', (keys_res) => {
-                if (keys_res.status == 200) {
-                  answers.spaceId = space_res.body.space.id
-                  answers.spaceDomain = space_res.body.space.domain.replace('https://', '')
-                    .replace('/', '')
-
-                  var tokens = keys_res.body.api_keys.filter((token) => {
-                    return token.access == 'theme'
-                  })
-
-                  answers.themeToken = tokens[0].token
-
-                  console.log(chalk.green('✓') + ' - Development Environment configured (./' + answers.name + '/config.js' + ')')
-                  
-                  lastStep()
-                } else {
-                  console.log(keys_res.body)
-                }
-              })
-            })
-          } else {
-            console.log('  Something went wrong, be sure that you inserted the right space id.')
-            console.log(space_res.body)
-          }
-        })
-
-      } else {
-
-        console.log('  Your project will be created now...')
-
-        api.post('spaces', {
-          // create_demo: true,
-          space: {
-            name: answers.name
-          }
-        }, (space_res) => {
-          if (space_res.status == 200) {
-            console.log(chalk.green('✓') + ' - Space ' + answers.name + ' has been created in Storyblok')
-            console.log(chalk.green('✓') + ' - Story "home" has been created in your Space')
-
-            api.post('tokens', {}, (tokens_res) => {
-              console.log(chalk.green('✓') + ' - Configuration for your Space was loaded')
-              answers.loginToken = tokens_res.body.key
-
-              api.setSpaceId(space_res.body.space.id)
-              api.get('api_keys', (keys_res) => {
-                if (keys_res.status == 200) {
-                  answers.spaceId = space_res.body.space.id
-                  answers.spaceDomain = space_res.body.space.domain.replace('https://', '')
-                    .replace('/', '')
-
-                  var tokens = keys_res.body.api_keys.filter((token) => {
-                    return token.access == 'theme'
-                  })
-
-                  console.log(chalk.green('✓') + ' - Starting Storyblok in your browser')
-                  
-                  setTimeout(() => {
-                    opn('http://' + answers.spaceDomain + '/_quickstart?quickstart=' + answers.loginToken)
-                    process.exit(0)
-                  }, 2000)
-                                      
-                  
-                } else {
-                  console.log(keys_res.body)
-                }
-              })
-            })
-          } else {
-            console.log(space_res.body)
-          }
-        })
-      }
+      quickstart(api, answers, spaceId)
 
       break;
     case 'push-components':
@@ -511,7 +339,6 @@ inquirer.prompt(questions).then(function (answers) {
 
       break;
     default:
-      lastStep()
+      lastStep(answers)
   }
-
 })
